Refactored activity options into separate component- provenanceActivityManager.jsx
